✨FIltra agora também por data, simultanemanete com os demais filtros

+++ backend/src/main/java/com/group/backend/crawler/PortaisCrawler.java
@@ -52,7 +52,7 @@ public class PortaisCrawler {
 
                 Noticia noticia = new Noticia();
                 noticia.setUrl(url);
-                noticia.setPorId(portalId.intValue());
+                noticia.setPorId(portalId.longValue());
 
                 controllerCrawler.startCrawlWithSeed(new String[]{url}, noticia);
